List of changes in description
**SYMBOLS PAGE:**

Changed the base name of ulity icon layer. To see the change, go to the symbols page. Find "Icons / Utility Icon." I changed the layer name of the icon symbol to "Icon" it was previously the original symbol name "Icons / z_embedded..."

Renamed the mask layer on standard icon and custom icon symbols to make the purpose clearer. (From "Mask" to "Rounded-Corner Mask"). Renamed action icon mask layer as well.

Check symbol was missing on the symbol "Desktop Components / Visual Picker / z-embedded / selected." Fixed this with the correct utility check.

Changed the color layer names on all the icon symbols from "Color" to emoji-Icon. This will make it much easier to scan in the overrides panel when there are 4+ overrides available.

Changed the arboard backgrounds for standard and custom icon symbols. It will make it easier to scan the icon artboards on the symbols page, but will NOT impact how they are used, so long as the "Export background" option remains UNSELECTED.

**TYPE & ICONS PAGE**
Made the header maks on the type and icons artboard narrower, locked the artboard group to prevent changes

DOCUMENT TYPE STYLES
Changed the names to support grouping: Body/etc ... allows for faster selection when switching between similar styles.