updated icon symbols to include sizing
tweaked action, custom and standard icon symbols to now include the scale:
- xx-small, 30x30px
- x-small, 32x32px
- small, 40x40px
- medium, 48x48px
- large, 64x64

and added the utility icon scale at
- xx-small, 14x14px
- x-small, 16x16px
- small, 24x24px
- medium 32x32px
- large 48x48px

Also adjusted all of the artboards to have the layers within a group called "bundle", resizing settings are on each folder to be fixed width, fixed height. This way if they are embedded within other symbols, the icon symbols will not scale inapporpriately.